Terraform/Azure/apps/func-apps-v1/requirements.txt
azure-functions
azure-storage-blob
pyarrow
pandas
rapidfuzz==3.6.1

